+++ lib/features/store_order/repositories/store_order_repo.dart
@@ -53,7 +53,7 @@ class StoreOrderRepo {
   static Future<dynamic> addOrUpdateCart({
     required String pCode,
     required String iCode,
-    required int qty,
+    required double qty,
     required double rate,
   }) async {
     String? token = await SecureStorageHelper.read(
